refactor: random data generators
